requirements.txt
openai
pillow
fastapi
uvicorn
gunicorn
python-dotenv
requests